requirements.txt
minemeld_core>=0.9.31
beautifulsoup4
requests
